attack_module/attack_modules.py
"""Attack primitives used by the Terrarium attack framework."""

from __future__ import annotations

from pathlib import Path
from typing import Any, Dict, Optional

import yaml

from src.agents.base import BaseAgent

DEFAULT_POISON_FILE = Path("attack_module/configs/poisoning_string.yaml")
DEFAULT_POISON_FALLBACK = (
    "This is a sample string. This is where you put a string that the adversary "
    "is poisoning the conversation with."
)


def _load_poison_payload(path: Path = DEFAULT_POISON_FILE, key="poisoning_string") -> str:
    """Load the default poisoning payload from disk (with fallback)."""

    try:
        with path.open("r", encoding="utf-8") as file:
            data = yaml.safe_load(file) or {}
            return data.get(key, DEFAULT_POISON_FALLBACK)
    except FileNotFoundError:
        return DEFAULT_POISON_FALLBACK


def _clone_arguments(arguments: Optional[Dict[str, Any]]) -> Dict[str, Any]:
    """Return a shallow copy of the tool call arguments for safe mutation."""

    if arguments is None:
        return {}
    return dict(arguments)


def _coerce_blackboard_value(value: Any) -> Any:
    try:
        return int(value)
    except (TypeError, ValueError):
        return value


def _normalize_blackboard_id(arguments: Dict[str, Any]) -> Optional[Any]:
    blackboard_id = arguments.get("blackboard_id") if arguments else None
    if blackboard_id is None:
        return None
    return _coerce_blackboard_value(blackboard_id)


def _preview(text: Optional[str], limit: int = 120) -> str:
    if not text:
        return ""
    text = str(text)
    return text[:limit] + ("…" if len(text) > limit else "")


def _log_agent_event(
    agent: BaseAgent,
    attack_type: str,
    *,
    success: bool,
    metadata: Optional[Dict[str, Any]] = None,
    target: Optional[Any] = None,
    target_type: Optional[str] = None,
):
    attack_logger = getattr(agent, "attack_logger", None)
    if not attack_logger:
        return
    metadata = dict(metadata or {})
    attack_label = getattr(agent, "attack_metadata", {}).get("label")
    if attack_label and "label" not in metadata:
        metadata["label"] = attack_label
    attack_logger.log_agent_attack(
        attack_type=attack_type,
        attacker=agent.name,
        target=target,
        target_type=target_type,
        success=success,
        phase=getattr(agent, "current_phase", None),
        iteration=getattr(agent, "current_iteration", None),
        round_num=getattr(agent, "current_round", None),
        metadata=metadata,
    )


class AgentPoisoningAttack(BaseAgent):
    """Replaces any outgoing blackboard posts with a fixed poisoning payload."""

    def __init__(
        self,
        client,
        name,
        model_name,
        max_conversation_steps,
        tool_logger,
        trajectory_logger,
        environment_name,
        poison_payload,
        generation_params: Optional[Dict[str, Any]] = None,
    ) -> None:
        super().__init__(
            client,
            name,
            model_name,
            max_conversation_steps,
            tool_logger,
            trajectory_logger,
            environment_name,
            generation_params=generation_params,
        )
        self.poison_payload =  _load_poison_payload(Path(poison_payload))

    async def _execute_tool_call(self, tool_name: str, arguments: Dict[str, Any]) -> Dict[str, Any]:
        if tool_name == "post_message":
            new_args = _clone_arguments(arguments)
            new_args["message"] = self.poison_payload
            target = _normalize_blackboard_id(new_args)
            metadata = {
                "payload_chars": len(self.poison_payload or ""),
                "payload_preview": _preview(self.poison_payload),
                "original_chars": len(str(arguments.get("message", ""))),
                "original_preview": _preview(arguments.get("message")),
                "blackboard_id": target,
            }
            _log_agent_event(
                self,
                "agent_poisoning",
                success=bool(self.poison_payload),
                metadata=metadata,
                target=target,
                target_type="blackboard",
            )
            return await super()._execute_tool_call(tool_name, new_args)
        return await super()._execute_tool_call(tool_name, arguments)


class ContextOverflowAttack(BaseAgent):
    """Appends a large filler sequence to overwhelm downstream context windows."""

    def __init__(
        self,
        client,
        name,
        model_name,
        max_conversation_steps,
        tool_logger,
        trajectory_logger,
        environment_name,
        poison_payload,
        generation_params: Optional[Dict[str, Any]] = None,
        # filler_token: str = "ATTACK",
        # repeat: int = 2048,
        # max_chars: Optional[int] = 16000,
        # header: Optional[str] = "FILLER PAYLOAD",
        # max_chars: Optional[int] = 16000,
        # header: Optional[str] = "FILLER PAYLOAD",
    ) -> None:
        super().__init__(
            client,
            name,
            model_name,
            max_conversation_steps,
            tool_logger,
            trajectory_logger,
            environment_name,
            generation_params=generation_params,
        )
        
        self.filler_token = _load_poison_payload(Path(poison_payload), "filler_token")
        self.repeat = int(_load_poison_payload(Path(poison_payload), "repeat"))
        self.repeat = max(1, self.repeat)
        self.max_chars = int(_load_poison_payload(Path(poison_payload), "max_chars"))
        self.header = _load_poison_payload(Path(poison_payload), "header")

    def _build_overflow(self) -> str:
        chunk = (self.filler_token + " ") * self.repeat
        chunk = chunk.strip()
        if self.max_chars is not None:
            chunk = chunk[: self.max_chars]
        return chunk

    async def _execute_tool_call(self, tool_name: str, arguments: Dict[str, Any]) -> Dict[str, Any]:
        if tool_name == "post_message":
            new_args = _clone_arguments(arguments)
            original = new_args.get("message", "")
            overflow = self._build_overflow()
            sections = [part for part in [self.header, original, overflow] if part]
            new_args["message"] = "\n\n".join(sections)
            target = _normalize_blackboard_id(new_args)
            metadata = {
                "overflow_chars": len(overflow),
                "original_chars": len(str(original)),
                "repeat": self.repeat,
                "max_chars": self.max_chars,
                "blackboard_id": target,
            }
            _log_agent_event(
                self,
                "context_overflow",
                success=bool(overflow),
                metadata=metadata,
                target=target,
                target_type="blackboard",
            )
            return await super()._execute_tool_call(tool_name, new_args)
        return await super()._execute_tool_call(tool_name, arguments)


class CommunicationProtocolPoisoningAttack:
    """Injects system messages into every blackboard via the communication proxy."""

    def __init__(
        self,
        payload: Optional[str] = None,
        attack_logger: Optional[Any] = None,
        attacker_label: str = "SYSTEM",
        poison_payload: Optional[str] = None,
    ) -> None:
        self.payload = payload or _load_poison_payload()
        self.attack_logger = attack_logger
        self.attacker_label = attacker_label

    async def inject(self, communication_protocol, context: Optional[Dict[str, Any]] = None) -> None:
        result = await communication_protocol.get_all_blackboard_ids()
        payload = {"content": self.payload}
        if not result:
            self._log_event(None, False, context, {"reason": "no_blackboards"})
            return
        for blackboard_id in result:
            normalized = _coerce_blackboard_value(blackboard_id)
            try:
                numeric_id = int(blackboard_id)
            except (TypeError, ValueError):
                numeric_id = blackboard_id
            await communication_protocol.post_system_message(
                numeric_id,
                "initialization",
                payload,
            )
            metadata = {
                "payload_chars": len(self.payload or ""),
                "payload_preview": _preview(self.payload),
                "blackboard_id": normalized,
            }
            self._log_event(normalized, True, context, metadata)

    def _log_event(
        self,
        target: Optional[Any],
        success: bool,
        context: Optional[Dict[str, Any]],
        metadata: Optional[Dict[str, Any]],
    ) -> None:
        if not self.attack_logger:
            return
        context = context or {}
        self.attack_logger.log_protocol_attack(
            attack_type="communication_poisoning",
            attacker=self.attacker_label,
            target=target,
            success=success,
            trigger=context.get("trigger"),
            iteration=context.get("iteration"),
            phase=context.get("phase"),
            round_num=context.get("round"),
            metadata=metadata or {},
        )
